move helper function to dtypes.py
